Type fix for code generation (fixing regressions visible on 32-bit big-endian hosts), partial fix for bug #3539.
git-svn-id: http://svn.code.sourceforge.net/p/sdcc/code/trunk/sdcc@13805 4a8a32a2-be11-0410-ad9d-d568d2c75423

+++ src/stm8/gen.c
@@ -8200,7 +8200,7 @@ genPointerGet (const iCode *ic)
   operand *left = IC_LEFT (ic);
   operand *right = IC_RIGHT (ic);
   int size, i;
-  long offset;
+  int offset;
   bool use_y;
   bool pushed_x = false;
   bool pushed_a = false;
